requirements.txt
pandas 
numpy
mysql-connector-python
pymysql
python-dotenv
scikit-learn
seaborn
matplotlib
catboost
#-e .